Fix the generic interface to work with Apache 2.4's mod_proxy_fcgi.
This new module has no support for "nph-" style applications, and is getting
confused by the generic interface's NPH behavior.

The correct fix would be to the Apache module, but absent the clue to do it,
the only thing I can do is hack the NPH behavior out of the generic interface.

It helps (or not) that I do not know why it is NPH in the first place. There
are some comments in the transport modules to the effect that their author
may have been the victim of some confusion on the subject. (Paraphrasing:
"We build an HTTP/1.x protocol identifier because IIS does NPH when it gets
one back; we ought to send Status: [i.e. non-NPH] instead but that breaks
Apache [no wonder, in a file named nph-*].")

My own, admittedly limited, use cases for the generic interface are fixed
by this change; there are probably some others out there that break instead.

Final editorializing: What on Earth is NPH good for anyway? There are only
two things I think may be plausible: One is sending invalid headers (which
you shouldn't do), the other is streaming responses (which OTRS doesn't do).