Support providing secrets via the process environment.
The FCGI module replaces the process environment with the request
environment received from the web server. This is normally desired
because it is the way CGI works as well, and because any CGI program
can be turned into a FastCGI program by wrapping its main entry point
into the FastCGI request loop, it is nominally correct.

When the FastCGI application is controlled by the web server, such
as with Apache's mod_fcgid etc., the web server is also responsible
for providing any environment variables outside the CGI request
environment that the application needs.

In this case, however, I do not have any plans to change the existing
arrangement where the application is controlled by daemontools
(or, in any case, externally to the web server), so there is no
reason not to keep parameters like credentials out of the web server
configuration and provide them in the process environment instead.
To do this, these variables are now copied from the process environment
before the FastCGI request loop and merged back into the request
environment for each request.