Reinstate the SendmailModule::CMD sysconfig option.
This reverts the removal of the option in 6.3.2 (6e4c048d) on security
grounds (OTRS OSA-2022-03, CVE-2021-36100) because the advisory says that
a "[s]pecially crafted string in OTRS system configuration can allow
the execution of any system command". This seems to me like the problem
exists on the other side of this airtight hatchway, i.e. to get the
"specially crafted string" _into_ sysconfig, you have to be an admin,
at which point you can just put in the command you want run.