Ticket#2023060210000281: Fix saving customer users
There is a bug in the code that processes form data from the customer
user edit form. It apparently looks up the record by the login field
regardless of what the configuration says. It should use the same
field (CustomerKey in Config.pm) that is put into the form by the
template.

Fix by changing the CustomerKey from 'email' to 'login'; the values
are unique as well.